population module renamed to explore module
