+++ src/components/analysis/filters/active-filters/index.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 
-import { selectPopulationProps } from 'modules/population/selectors';
+import { selectPopulationProps } from 'modules/explore/selectors';
 
 import Component from './component';
