+++ src/components/data-detail/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 
 import Component from './component';
-import { selectPopulationDetailProps } from 'modules/population/selectors';
+import { selectPopulationDetailProps } from 'modules/explore/selectors';
 
 export default connect(selectPopulationDetailProps)(Component);
