+++ src/components/filters/clear-filters/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 
-import { selectPopulationProps } from 'modules/population/selectors';
-import { resetFilters } from 'modules/population/actions';
+import { selectPopulationProps } from 'modules/explore/selectors';
+import { resetFilters } from 'modules/explore/actions';
 
 import Component from './component';
