+++ src/pages/explore/detail/map/index.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 
 import { setRouter } from 'modules/router/actions';
 import { selectMapProps } from 'modules/map/selectors';
-import { selectPopulationDetailProps } from 'modules/population/selectors';
+import { selectPopulationDetailProps } from 'modules/explore/selectors';
 
 import Component from './component';
