+++ src/pages/explore/overview/map/index.js
@@ -1,9 +1,9 @@
 import { connect } from 'react-redux';
 
 import { setPopUp, setLocation } from 'modules/map/actions';
-import { setPopulationsByLocation } from 'modules/population/actions';
+import { setPopulationsByLocation } from 'modules/explore/actions';
 import { selectMapProps } from 'modules/map/selectors';
-import { selectPopulationDetailProps, selectPopulationProps } from 'modules/population/selectors';
+import { selectPopulationDetailProps, selectPopulationProps } from 'modules/explore/selectors';
 
 
 import Component from './component';
